feat: Add support for model downloading with metadata management
