+++ manager/src/main.rs
@@ -162,6 +162,7 @@ impl AppConfig {
     }
 }
 
+pub mod downloader;
 pub mod auth;
 pub mod setup;
 
@@ -420,7 +421,7 @@ pub(crate) async fn trigger_download(
     let shutdown_rx = state.shutdown_tx.subscribe();
 
     let task = tokio::spawn(async move {
-        perform_model_download(state_clone, id_clone, repo, filename, shutdown_rx).await;
+        downloader::perform_model_download(state_clone, id_clone, repo, filename, shutdown_rx).await;
     });
 
     state
@@ -467,484 +468,6 @@ pub(crate) async fn pause_download(
         (StatusCode::OK, "Download paused").into_response()
     } else {
         (StatusCode::NOT_FOUND, "No active download to pause").into_response()
-    }
-}
-
-struct DownloadCleanupGuard {
-    state: Arc<AppState>,
-    id: String,
-}
-
-impl Drop for DownloadCleanupGuard {
-    fn drop(&mut self) {
-        self.state
-            .active_downloads
-            .lock()
-            .unwrap_or_else(|e| e.into_inner())
-            .remove(&self.id);
-        self.state
-            .download_tasks
-            .lock()
-            .unwrap_or_else(|e| e.into_inner())
-            .remove(&self.id);
-    }
-}
-
-async fn perform_model_download(
-    state: Arc<AppState>,
-    id: String,
-    repo: String,
-    filename: String,
-    mut shutdown_rx: tokio::sync::broadcast::Receiver<()>,
-) {
-    let _guard = DownloadCleanupGuard {
-        state: state.clone(),
-        id: id.clone(),
-    };
-
-    let url = format!(
-        "{}/{}/resolve/main/{}",
-        state.config.hf_base_url.trim_end_matches('/'), repo, filename
-    );
-    let client = &state.reqwest_client;
-
-    let downloads_dir = &state.config.downloads_directory;
-    if let Err(e) = tokio::fs::create_dir_all(downloads_dir).await {
-        error!("Failed to create downloads directory for {}: {}", id, e);
-        return;
-    }
-    let file_path = std::path::Path::new(downloads_dir).join(&filename);
-    let tmp_file_path = {
-        let mut p = file_path.clone().into_os_string();
-        p.push(".tmp");
-        std::path::PathBuf::from(p)
-    };
-    let meta_file_path = {
-        let mut p = file_path.clone().into_os_string();
-        p.push(".meta");
-        std::path::PathBuf::from(p)
-    };
-
-    let mut existing_size = 0;
-    let mut expected_hash = None;
-
-    if let Ok(meta_str) = tokio::fs::read_to_string(&meta_file_path).await {
-        if let Ok(meta) = serde_json::from_str::<serde_json::Value>(&meta_str)
-            && let Some(bytes) = meta.get("downloaded_bytes").and_then(|v| v.as_u64())
-        {
-            existing_size = bytes;
-            if let Some(hash) = meta.get("expected_hash").and_then(|v| v.as_str()) {
-                expected_hash = Some(hash.to_string());
-            }
-        } else {
-            warn!(
-                "Corrupted or invalid metadata file found for {}. Discarding and restarting download.",
-                id
-            );
-            {
-                let mut dl = state
-                    .active_downloads
-                    .lock()
-                    .unwrap_or_else(|e| e.into_inner());
-                if let Some(status) = dl.get_mut(&id) {
-                    status.state = "Corrupted metadata. Restarting...".to_string();
-                }
-            }
-            let _ = tokio::fs::remove_file(&meta_file_path).await;
-            let _ = tokio::fs::remove_file(&tmp_file_path).await;
-            tokio::time::sleep(std::time::Duration::from_secs(3)).await;
-        }
-    }
-
-    if existing_size > 0 {
-        if let Ok(file) = tokio::fs::OpenOptions::new()
-            .write(true)
-            .open(&tmp_file_path)
-            .await
-            && let Ok(metadata) = file.metadata().await
-        {
-            let actual_size = metadata.len();
-            if actual_size < existing_size {
-                existing_size = actual_size;
-            } else if actual_size > existing_size {
-                let _ = file.set_len(existing_size).await;
-            }
-        } else {
-            existing_size = 0;
-        }
-    }
-
-    {
-        let mut dl = state
-            .active_downloads
-            .lock()
-            .unwrap_or_else(|e| e.into_inner());
-        if let Some(status) = dl.get_mut(&id) {
-            status.bytes_transferred = existing_size;
-            status.state = "Connecting...".to_string();
-        }
-    }
-
-    let mut hasher = Sha256::new();
-
-    let mut req = client.get(&url);
-    if existing_size > 0 {
-        req = req.header(reqwest::header::RANGE, format!("bytes={}-", existing_size));
-    }
-
-    let mut res = match req.send().await {
-        Ok(r) => {
-            if r.status() == reqwest::StatusCode::RANGE_NOT_SATISFIABLE {
-                existing_size = 0;
-                hasher = Sha256::new();
-                match client.get(&url).send().await {
-                    Ok(r2) => {
-                        if !r2.status().is_success() {
-                            error!("Download failed for {} with status: {}", id, r2.status());
-                            return;
-                        }
-                        r2
-                    }
-                    Err(e) => {
-                        error!("Download failed for {}: {}", id, e);
-                        return;
-                    }
-                }
-            } else if !r.status().is_success() {
-                error!("Download failed for {} with status: {}", id, r.status());
-                return;
-            } else {
-                r
-            }
-        }
-        Err(e) => {
-            error!("Download failed for {}: {}", id, e);
-            return;
-        }
-    };
-
-        let mut new_etag = None;
-    if let Some(etag) = res
-        .headers()
-        .get("X-Linked-Etag")
-        .or_else(|| res.headers().get("ETag"))
-    {
-        let etag_str = etag.to_str().unwrap_or("").trim_matches('"');
-        let clean_etag = etag_str
-            .strip_prefix("W/")
-            .unwrap_or(etag_str)
-            .trim_matches('"');
-        if clean_etag.len() == 64 && clean_etag.chars().all(|c| c.is_ascii_hexdigit()) {
-                new_etag = Some(clean_etag.to_string());
-        }
-    }
-
-        if existing_size > 0 && expected_hash.is_some() && new_etag.is_some() && expected_hash != new_etag {
-            warn!(
-                "ETag mismatch for {} (expected {:?}, got {:?}). Restarting download.",
-                id, expected_hash, new_etag
-            );
-            existing_size = 0;
-            hasher = Sha256::new();
-            expected_hash = new_etag;
-
-            res = match client.get(&url).send().await {
-                Ok(r) => {
-                    if !r.status().is_success() {
-                        error!("Download failed for {} on restart with status: {}", id, r.status());
-                        return;
-                    }
-                    r
-                }
-                Err(e) => {
-                    error!("Download failed for {} on restart: {}", id, e);
-                    return;
-                }
-            };
-        } else if new_etag.is_some() {
-            expected_hash = new_etag;
-        }
-
-        let is_partial = res.status() == reqwest::StatusCode::PARTIAL_CONTENT;
-        if !is_partial && existing_size > 0 {
-            info!(
-                "Server did not return partial content, restarting download for {}",
-                id
-            );
-            existing_size = 0;
-            hasher = Sha256::new();
-        }
-
-    let total_size = if is_partial {
-        existing_size + res.content_length().unwrap_or(0)
-    } else {
-        res.content_length().unwrap_or(0)
-    };
-
-    // Pre-initialize metadata
-    let _ = tokio::fs::write(
-        &meta_file_path,
-        serde_json::json!({
-            "downloaded_bytes": existing_size,
-            "expected_hash": expected_hash
-        })
-        .to_string(),
-    )
-    .await;
-
-    {
-        let mut dl = state
-            .active_downloads
-            .lock()
-            .unwrap_or_else(|e| e.into_inner());
-        if let Some(status) = dl.get_mut(&id) {
-            status.total_bytes = total_size;
-            status.bytes_transferred = existing_size; // SHOW PROGRESS IMMEDIATELY
-            status.state = "Downloading...".to_string();
-        }
-    }
-
-    let mut file = match if is_partial && existing_size > 0 {
-        tokio::fs::OpenOptions::new()
-            .append(true)
-            .open(&tmp_file_path)
-            .await
-    } else {
-        tokio::fs::File::create(&tmp_file_path).await
-    } {
-        Ok(f) => f,
-        Err(e) => {
-            error!("Failed to open/create file for {}: {}", id, e);
-            return;
-        }
-    };
-
-    let start_time = std::time::Instant::now();
-    let mut downloaded: u64 = existing_size;
-    let mut stream_error = false;
-    let mut last_meta_save = std::time::Instant::now();
-    let mut last_ui_update = std::time::Instant::now();
-    let mut bytes_since_last_ui_update = 0;
-
-    let update_interval_ms = 500;
-    let update_bytes_threshold = 1024 * 1024; // 1 MB
-
-    use tokio::io::AsyncWriteExt;
-
-    loop {
-        tokio::select! {
-            chunk_res = res.chunk() => {
-                let bytes = match chunk_res {
-                    Ok(Some(c)) => c,
-                    Ok(None) => break,
-                    Err(e) => {
-                        error!("Error while streaming {}: {}", id, e);
-                        stream_error = true;
-                        break;
-                    }
-                };
-
-                if existing_size == 0 {
-                    hasher.update(&bytes);
-                }
-                if let Err(e) = file.write_all(&bytes).await {
-                    error!("Failed to write to file for {}: {}", id, e);
-                    stream_error = true;
-                    break;
-                }
-                downloaded += bytes.len() as u64;
-                bytes_since_last_ui_update += bytes.len() as u64;
-
-                if last_ui_update.elapsed().as_millis() >= update_interval_ms
-                    || bytes_since_last_ui_update >= update_bytes_threshold
-                {
-                    let elapsed = start_time.elapsed().as_secs_f64();
-                    let speed = if elapsed > 0.0 {
-                        (downloaded.saturating_sub(existing_size)) as f64 / elapsed
-                    } else {
-                        0.0
-                    };
-
-                    {
-                        let mut dl = state
-                            .active_downloads
-                            .lock()
-                            .unwrap_or_else(|e| e.into_inner());
-                        if let Some(status) = dl.get_mut(&id) {
-                            status.bytes_transferred = downloaded;
-                            status.current_speed_bps = speed;
-                            status.state = "Downloading...".to_string();
-                        }
-                    }
-
-                    last_ui_update = std::time::Instant::now();
-                    bytes_since_last_ui_update = 0;
-                }
-
-                if last_meta_save.elapsed().as_secs() > 5 {
-                    let _ = tokio::fs::write(
-                        &meta_file_path,
-                        serde_json::json!({
-                            "downloaded_bytes": downloaded,
-                            "expected_hash": expected_hash
-                        })
-                        .to_string(),
-                    )
-                    .await;
-                    last_meta_save = std::time::Instant::now();
-                }
-            }
-            _ = shutdown_rx.recv() => {
-                info!("Shutdown signal received. Flushing metadata for {}...", id);
-                let _ = tokio::fs::write(
-                    &meta_file_path,
-                    serde_json::json!({
-                        "downloaded_bytes": downloaded,
-                        "expected_hash": expected_hash
-                    })
-                    .to_string(),
-                )
-                .await;
-                return;
-            }
-        }
-    }
-
-    // Explicitly drop the file handle so Windows allows us to rename/delete it
-    drop(file);
-
-    let mut hash_mismatch = false;
-
-    if !stream_error {
-        let final_hash_opt = if existing_size == 0 {
-            Some(hex::encode(hasher.finalize()))
-        } else {
-            {
-                let mut dl = state
-                    .active_downloads
-                    .lock()
-                    .unwrap_or_else(|e| e.into_inner());
-                if let Some(status) = dl.get_mut(&id) {
-                    status.state = "Verifying...".to_string();
-                }
-            }
-
-            let tmp_file_path_clone = tmp_file_path.clone();
-            let hash_result = tokio::task::spawn_blocking(move || -> std::io::Result<String> {
-                let mut h = Sha256::new();
-                use std::io::Read;
-                let mut f = std::fs::File::open(&tmp_file_path_clone)?;
-                let mut buf = vec![0u8; 4 * 1024 * 1024];
-                loop {
-                    let n = f.read(&mut buf)?;
-                    if n == 0 {
-                        break;
-                    }
-                    h.update(&buf[..n]);
-                }
-                Ok(hex::encode(h.finalize()))
-            })
-            .await;
-
-            match hash_result {
-                Ok(Ok(hash)) => Some(hash),
-                Ok(Err(e)) => {
-                    error!("Failed to read file for verification {}: {}", id, e);
-                    None
-                }
-                Err(e) => {
-                    error!("Hashing task panicked or was cancelled for {}: {}", id, e);
-                    None
-                }
-            }
-        };
-
-        if let Some(final_hash) = final_hash_opt {
-            if let Some(ref expected) = expected_hash {
-                if &final_hash != expected {
-                    error!(
-                        "Checksum mismatch for {}. Expected {}, got {}. Marking as corrupted.",
-                        id, expected, final_hash
-                    );
-                    hash_mismatch = true;
-                } else {
-                    info!("Checksum validated successfully for {}.", id);
-                }
-            } else {
-                info!(
-                    "No SHA-256 ETag found for {}. Saving. Computed: {}",
-                    id, final_hash
-                );
-            }
-        }
-    }
-
-    if stream_error {
-        let _ = tokio::fs::write(
-            &meta_file_path,
-            serde_json::json!({
-                "downloaded_bytes": downloaded,
-                "expected_hash": expected_hash
-            })
-            .to_string(),
-        )
-        .await;
-        info!(
-            "Network interrupted. Kept partial temp file for {} to resume later.",
-            id
-        );
-    } else {
-        let target_path = if hash_mismatch {
-            let mut p = file_path.clone().into_os_string();
-            p.push(".corrupted");
-            std::path::PathBuf::from(p)
-        } else {
-            file_path.clone()
-        };
-
-        let mut success = false;
-        if let Err(e) = tokio::fs::rename(&tmp_file_path, &target_path).await {
-            warn!(
-                "Failed to rename temp file for {} ({}). Falling back to copy...",
-                id, e
-            );
-            match tokio::fs::copy(&tmp_file_path, &target_path).await {
-                Ok(_) => {
-                    let _ = tokio::fs::remove_file(&tmp_file_path).await;
-                    success = true;
-                }
-                Err(copy_err) => {
-                    error!(
-                        "Failed to copy temp file to final file for {}: {}",
-                        id, copy_err
-                    );
-                }
-            }
-        } else {
-            success = true;
-        }
-
-        if success {
-            let _ = tokio::fs::remove_file(&meta_file_path).await;
-            if hash_mismatch {
-                error!(
-                    "Download finished but checksum failed for {}. Saved as {}. Please verify the file and rename it manually to {} if it is valid.",
-                    id,
-                    target_path.display(),
-                    filename
-                );
-                {
-                    let mut status = lock_status(&state.engine_status);
-                    status.corrupted_models.insert(id.clone());
-                }
-            } else {
-                info!("Finished downloading {}", id);
-                {
-                    let mut status = lock_status(&state.engine_status);
-                    status.downloaded_models.insert(id.clone());
-                }
-            }
-        }
     }
 }
